original assignment/task description
To speed up the process on writing paragraphs/sections for mandatory
assignments, I've added the original task description to simplify
things. To not spend to much time re-writing something that may not be
used, I just used Google Translate to get a rough translation. This is
just in-case I want to re-use the text later, but also ease the "mental
translation" when re-reading tasks and project goals.