Re-written Section 3
Section 3 is now fully re-written, the exception being the section in
relation to the machine learning library. This might be updated later
on, but seeing as the code for this is in English, it is not a priority
for the time being.
Also updated the web-section, so that they include the correct "tags" in
their description (e.g. course, ide, prog lang, etc.)