Introduction and Indexing
Written a more proper introduction that reflects the content of the
document.
Added revision history, and link to the repository the document exists
in (for those getting it to be able to see latest version if updates are
required). Bound the Index to the ToC, so that readers won't have to
scroll to last page to see it.